moved tests to appropriate folder
